// src/app/components/enums/color-enum.ts
export enum Color {
  BlueP = '#1eb4e9',
  BlueS = '#bbe8f8',
  BlueT = '#2a98c2',
  BlueF = 'hwb(196 73 % 3 %)',
  YellowP = '#fff0d2',
  YellowS = '#9b8930',
  PurpleP = '#ddcfe8',
  PurpleS = '#7660b4',
  PurpleT = 'hsla(256, 100 %, 91 %, 0.3)',
  GreenP = '#c3ecca',
  GreenS = ' #2f7468',
  GreenT = 'hsla(170, 78 %, 66 %, 0.3)',
  GreyP = '#dedede',
  GreyS = '#bbbbbb',
  Black = 'black',
  White = 'white',
  Red = '#F66F6F',
  Shadow = 'rgba(0, 0, 0, 0.16)',
}